perform initial routing refactor
