+++ src/ui/routes/package/[name]/deploy/+page.svelte
