+++ src/ui/routes/package/[name]/finished/+page.svelte
